implemented the layout component in gatsby-ssr.js
